In progress updates for unified translation system
