+++ .dockerignore
@@ -1,0 +1,10 @@
+.git
+safe
+realtime
+i18n
+safe_extras
+scripts
+docs
+extras
+resources
+venv
