+++ safe/common/utilities.py
@@ -18,6 +18,8 @@ from collections import OrderedDict
 # pylint: enable=W0611
 
 from safe.common.exceptions import VerificationError
+from safe.utilities.i18n import locale
+
 
 import logging
 LOGGER = logging.getLogger('InaSAFE')
@@ -319,19 +321,19 @@ def format_int(x):
         return x
 
     # Quick solution for the moment
-    if lang == 'id':
+    if locale() == 'id':
         # Replace commas with dots
         s = s.replace(',', '.')
     return s
 
 
-def round_thousand(my_int):
+def round_thousand(value):
     """Round an integer to the nearest thousand if my_int
     is more than a thousand
     """
-    if my_int > 1000:
-        my_int = my_int // 1000 * 1000
-    return my_int
+    if value > 1000:
+        value = value // 1000 * 1000
+    return value
 
 
 def humanize_min_max(min_value, max_value, interval):
@@ -339,15 +341,19 @@ def humanize_min_max(min_value, max_value, interval):
     If the range between the max and min is less than one, the original
     value will be returned.
 
-    Args:
-        * min_value
-        * max_value
-        * interval - (float): the interval between classes in the the
-            class list where the results will be used.
+    :param min_value: Minimum value
+    :type min_value: int, float
 
-    Returns:
-        A two-tuple consisting of a string for min_value and a string for
+    :param max_value: Maximim value
+    :type max_value: int, float
+
+    :param interval: The interval between classes in the
+            class list where the results will be used.
+    :type interval: float, int
+
+    :returns: A two-tuple consisting of a string for min_value and a string for
             max_value.
+    :rtype: tuple
 
     """
     current_interval = max_value - min_value
@@ -363,7 +369,7 @@ def humanize_min_max(min_value, max_value, interval):
     return humanize_min_value, humanize_max_value
 
 
-def format_decimal(interval, my_number):
+def format_decimal(interval, value):
     """Return formatted decimal according to interval decimal place
     For example:
     interval = 0.33 (two decimal places)
@@ -373,17 +379,17 @@ def format_decimal(interval, my_number):
     If my_number is an integer return as is
     """
     interval = get_significant_decimal(interval)
-    if isinstance(interval, Integral) or isinstance(my_number, Integral):
-        return format_int(int(my_number))
+    if isinstance(interval, Integral) or isinstance(value, Integral):
+        return format_int(int(value))
     if interval != interval:
         # nan
-        return str(my_number)
-    if my_number != my_number:
+        return str(value)
+    if value != value:
         # nan
-        return str(my_number)
+        return str(value)
     decimal_places = len(str(interval).split('.')[1])
-    my_number_int = str(my_number).split('.')[0]
-    my_number_decimal = str(my_number).split('.')[1][:decimal_places]
+    my_number_int = str(value).split('.')[0]
+    my_number_decimal = str(value).split('.')[1][:decimal_places]
     if len(set(my_number_decimal)) == 1 and my_number_decimal[-1] == '0':
         return my_number_int
     return (format_int(int(my_number_int)) + get_decimal_separator() +
